better code

+++ Laba1/CSVFileReader.cpp
@@ -42,8 +42,8 @@ std::vector<std::vector<double>> CSVFileReader::read(const std::string& fileName
         // Чтение строк данных матрицы
         for (int i = 0; i < rowCount && std::getline(inputFile, line); ++i) {
             int col = 0;
-            size_t start = 0;
-            size_t end;
+            int start = 0;
+            int end;
 
             while ((end = line.find(';', start)) != std::string::npos) {
                 if (col >= columnCount) {
